+++ modules/polls/polls.uninstall.php
@@ -23,18 +23,17 @@ if (!isset($res)) {
 	$res = '';
 }
 
-global $db_auth;
+global $db_auth, $cfg;
 if (isset($db_auth)) {
 	$sql = sed_sql_query("DELETE FROM $db_auth WHERE auth_code='polls'");
 	$res .= "Deleted polls auth entries: " . sed_sql_affectedrows() . "<br />";
 }
 
-/* Tables (polls, polls_options, polls_voters) are left in place to preserve data.
-   To drop them on uninstall, uncomment the lines below:
-
-sed_sql_query("DROP TABLE IF EXISTS " . $cfg['sqldbprefix'] . "polls_voters");
-sed_sql_query("DROP TABLE IF EXISTS " . $cfg['sqldbprefix'] . "polls_options");
-sed_sql_query("DROP TABLE IF EXISTS " . $cfg['sqldbprefix'] . "polls");
-*/
-
-$res .= "Polls module uninstalled. Poll tables preserved.<br />";
+if (!empty($sed_uninstall_drop_tables)) {
+	sed_sql_query("DROP TABLE IF EXISTS " . $cfg['sqldbprefix'] . "polls_voters");
+	sed_sql_query("DROP TABLE IF EXISTS " . $cfg['sqldbprefix'] . "polls_options");
+	sed_sql_query("DROP TABLE IF EXISTS " . $cfg['sqldbprefix'] . "polls");
+	$res .= "Poll tables dropped.<br />";
+} else {
+	$res .= "Polls module uninstalled. Poll tables preserved.<br />";
+}
